[docs] Try CodeFund over Carbon (#12484)
