+++ docs/src/modules/components/MarkdownDocs.js
@@ -8,7 +8,6 @@ import MarkdownElement from '@material-ui/docs/MarkdownElement';
 import Head from 'docs/src/modules/components/Head';
 import AppContent from 'docs/src/modules/components/AppContent';
 import Demo from 'docs/src/modules/components/Demo';
-import Carbon from 'docs/src/modules/components/Carbon';
 import AppFrame from 'docs/src/modules/components/AppFrame';
 import AppTableOfContents from 'docs/src/modules/components/AppTableOfContents';
 import {
@@ -38,7 +37,7 @@ const demoRegexp = /^"demo": "(.*)"/;
 const SOURCE_CODE_ROOT_URL = 'https://github.com/mui-org/material-ui/tree/master';
 
 function MarkdownDocs(props, context) {
-  const { classes, demos, disableCarbon, markdown, markdownLocation: markdownLocationProp } = props;
+  const { classes, demos, disableAd, markdown, markdownLocation: markdownLocationProp } = props;
   const contents = getContents(markdown);
   const headers = getHeaders(markdown);
 
@@ -78,14 +77,13 @@ ${headers.components
         title={`${headers.title || getTitle(markdown)} - Material-UI`}
         description={getDescription(markdown)}
       />
-      <AppTableOfContents contents={contents} />
+      <AppTableOfContents contents={contents} disableAd={disableAd} />
       <AppContent className={classes.root}>
         <div className={classes.header}>
           <Button component="a" href={`${SOURCE_CODE_ROOT_URL}${markdownLocation}`}>
             {'Edit this page'}
           </Button>
         </div>
-        {disableCarbon ? null : <Carbon key={markdownLocation} />}
         {contents.map((content, index) => {
           const match = content.match(demoRegexp);
 
@@ -118,7 +116,7 @@ ${headers.components
 MarkdownDocs.propTypes = {
   classes: PropTypes.object.isRequired,
   demos: PropTypes.object,
-  disableCarbon: PropTypes.bool,
+  disableAd: PropTypes.bool,
   markdown: PropTypes.string.isRequired,
   // You can define the direction location of the markdown file.
   // Otherwise, we try to determine it with an heuristic.
@@ -126,7 +124,7 @@ MarkdownDocs.propTypes = {
 };
 
 MarkdownDocs.defaultProps = {
-  disableCarbon: false,
+  disableAd: false,
 };
 
 MarkdownDocs.contextTypes = {
