+++ pages/_document.js
@@ -60,13 +60,13 @@ class MyDocument extends Document {
         <body>
           <Main />
           {/* Global Site Tag (gtag.js) - Google Analytics */}
-          <script async src={`https://www.googletagmanager.com/gtag/js?id=${config.google.id}`} />
+          <script async src="https://www.google-analytics.com/analytics.js" />
           <script
             // eslint-disable-next-line react/no-danger
             dangerouslySetInnerHTML={{
               __html: `
-window.dataLayer = window.dataLayer || [];function gtag(){dataLayer.push(arguments)};
-gtag('js', new Date());
+window.ga=window.ga||function(){(ga.q=ga.q||[]).push(arguments)};ga.l=+new Date;
+ga('create', '${config.google.id}', 'material-ui.com');
               `,
             }}
           />
